rename files to be less verbose
